Move bounded writer into templates crate
